requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain
langchain_google_genai
streamlit
python-dotenv
langchain_community
PyPDF2